requirements.txt
discord.py==2.3.1
vt-py==0.17.5
pyfiglet==0.8.post1
nest-asyncio==1.5.7
emailrep==0.0.5
ipaddress==1.0.23
requests==2.31.0
shodan==1.29.1